这里我重构了整个项目，已解决AI的库与原本生态的版本冲突问题同时以实现简单的RAG检索增强，识别用户意图

+++ src/main/java/com/hmdp/service/impl/FollowServiceImpl.java
@@ -14,7 +14,7 @@ import com.hmdp.utils.UserHolder;
 import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.stereotype.Service;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 import java.util.Collections;
 import java.util.List;
 import java.util.Set;
@@ -63,7 +63,7 @@ public class FollowServiceImpl extends ServiceImpl<FollowMapper, Follow> impleme
     @Override
     public Result isFollow(Long followId) {
         Long userId = UserHolder.getUser().getId();
-        Integer count = query().eq("user_id", userId).eq("follow_user_id", followId).count();
+        Long count = query().eq("user_id", userId).eq("follow_user_id", followId).count();
         return Result.ok(count>0);
     }
