+++ src/main/java/com/hmdp/service/impl/VoucherOrderServiceImpl.java
@@ -19,8 +19,8 @@ import org.springframework.data.redis.core.script.DefaultRedisScript;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.annotation.PostConstruct;
-import javax.annotation.Resource;
+
+import jakarta.annotation.Resource;
 import java.util.Collections;
 import java.util.concurrent.*;
 
@@ -183,7 +183,7 @@ public class VoucherOrderServiceImpl extends ServiceImpl<VoucherOrderMapper, Vou
 //        Long userId = UserHolder.getUser().getId();
         Long userId = voucherOrder.getUserId();
         //查询订单
-        Integer count = query().eq("user_id", userId).eq("voucher_id", voucherOrder.getVoucherId()).count();
+        Long count = query().eq("user_id", userId).eq("voucher_id", voucherOrder.getVoucherId()).count();
         if (count>0)
             log.error("用户已经购买过一次！");
